+++ yacman/const.py
@@ -1,0 +1,15 @@
+"""
+Constant variables for yacman package
+"""
+
+FILEPATH_KEY = "_file_path"
+RO_KEY = "_ro"
+USE_LOCKS_KEY = "_locks"
+ORI_STATE_KEY = "_ori_state"
+WAIT_MAX_KEY = "_wait_time"
+
+ATTR_KEYS = (USE_LOCKS_KEY, FILEPATH_KEY, RO_KEY, ORI_STATE_KEY, WAIT_MAX_KEY)
+
+LOCK_PREFIX = "lock."
+DEFAULT_RO = False
+DEFAULT_WAIT_TIME = 10