+++ docs/changelog.md
@@ -2,6 +2,15 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [0.6.5] -- 2019-12-02
+
+### Added
+- context manager functionality to `YacAttMap` class
+
+### Changed
+- method name: `unlock` to `make_readonly`
+- `make_writable` behavior: it re-reads the source file now
+
 ## [0.6.4] -- 2019-11-04
 
 ### Added
@@ -20,7 +29,6 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - in `select_config` always use default config file path when no valid path is determined 
-
 
 ## [0.6.1] -- 2019-10-08